+++ bot/src/context/context.interface.ts
@@ -1,0 +1,7 @@
+import { Context } from "telegraf";
+
+export interface SessionDataI {}
+
+export interface BotContextI extends Context {
+  session: SessionDataI;
+}